+++ requirements.txt
@@ -9,3 +9,4 @@ gunicorn==20.1.0
 psycopg2==2.9.2
 snapshottest==0.6.0
 selenium==4.1.0
+nose==1.3.7
