+++ server/lib/stats_calculator.py
@@ -1,4 +1,4 @@
-from models import db, Stats, Player
+from server.models import db, Stats, Player
 
 class StatsCalculator:
     def __init__(self, game):
