+++ server/tests/snapshots/snap_backend_tests.py
@@ -1,0 +1,3518 @@
+# -*- coding: utf-8 -*-
+# snapshottest: v1 - https://goo.gl/zC4yUc
+from __future__ import unicode_literals
+
+from snapshottest import Snapshot
+
+
+snapshots = Snapshot()
+
+snapshots['StatsSnapshotTests::test_basic_point 1'] = {
+    'stats': {
+        'Heather McCabe': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Jamie Wildgen': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 1,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 10000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 10000,
+            'second_assists': 1,
+            'team': 'Soho',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Karen Kavanagh': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Kevin Barford': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Laura Knowles': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 1000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 1000,
+            'second_assists': 0,
+            'team': 'Soho',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Melissa Jess': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 1000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 1000,
+            'second_assists': 0,
+            'team': 'Soho',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Michael Colantonio': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 1,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 1,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 11500,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 11500,
+            'second_assists': 0,
+            'team': 'Soho',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Owen Lumley': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 1,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Sebastien Belanger': {
+            'assists': 1,
+            'callahan': 0,
+            'catches': 1,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 12000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 12000,
+            'second_assists': 0,
+            'team': 'Soho',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Sina Dee': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 1500,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 1500,
+            'second_assists': 0,
+            'team': 'Soho',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Stephen Close': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Wing-Leung Chan': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        }
+    },
+    'week': 0
+}
+
+snapshots['StatsSnapshotTests::test_callahan 1'] = {
+    'stats': {
+        'Ashlin Kelly': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 1,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 1,
+            'o_points_for': 0,
+            'pay': -4500,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': -4500,
+            'second_assists': 0,
+            'team': 'Kells Angels Bicycle Club',
+            'threw_drops': 0,
+            'throw_aways': 1,
+            'total_efficiency': 0.0
+        },
+        'Benjamin Piper': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 1,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'Kells Angels Bicycle Club',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Brent Burton': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 1.0,
+            'd_points_against': 0,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 2000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 2000,
+            'second_assists': 0,
+            'team': '99 Problems',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Brian Kells': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 1,
+            'o_points_for': 0,
+            'pay': 500,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 500,
+            'second_assists': 0,
+            'team': 'Kells Angels Bicycle Club',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Carrie-Anne Whyte': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 1,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'Kells Angels Bicycle Club',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Christopher Keates': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 1.0,
+            'd_points_against': 0,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 2000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 2000,
+            'second_assists': 0,
+            'team': '99 Problems',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Jessie Robinson': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 1.0,
+            'd_points_against': 0,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 2000,
+            'points_played': 1,
+            'pulls': 1,
+            'salary_per_point': 2000,
+            'second_assists': 0,
+            'team': '99 Problems',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Jim Robinson': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 1,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'Kells Angels Bicycle Club',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Justine Price': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 1.0,
+            'd_points_against': 0,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 2000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 2000,
+            'second_assists': 0,
+            'team': '99 Problems',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Kevin Hughes(S)': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 1,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 1,
+            'o_points_for': 0,
+            'pay': 1000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 1000,
+            'second_assists': 0,
+            'team': 'Kells Angels Bicycle Club',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Krys Kudakiewicz': {
+            'assists': 0,
+            'callahan': 1,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 1,
+            'd_efficiency': 1.0,
+            'd_points_against': 0,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 1,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 20000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 20000,
+            'second_assists': 0,
+            'team': '99 Problems',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Rob Ives': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 1.0,
+            'd_points_against': 0,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 2000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 2000,
+            'second_assists': 0,
+            'team': '99 Problems',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        }
+    },
+    'week': 0
+}
+
+snapshots['StatsSnapshotTests::test_catch_d 1'] = {
+    'stats': {
+        'Brian Kells': {
+            'assists': 1,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 1,
+            'd_blocks': 1,
+            'd_efficiency': 1.0,
+            'd_points_against': 0,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 20500,
+            'points_played': 1,
+            'pulls': 1,
+            'salary_per_point': 20500,
+            'second_assists': 0,
+            'team': 'Attack',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Christopher Keates': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 1,
+            'o_points_for': 0,
+            'pay': 500,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 500,
+            'second_assists': 0,
+            'team': 'Huck and Hope School of Handling (Hope)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Hannah Dawson': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 1.0,
+            'd_points_against': 0,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 2000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 2000,
+            'second_assists': 0,
+            'team': 'Attack',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Jonathan Champagne': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 1.0,
+            'd_points_against': 0,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 2000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 2000,
+            'second_assists': 0,
+            'team': 'Attack',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Kevin Barford': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 1,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 1,
+            'o_points_for': 0,
+            'pay': -4500,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': -4500,
+            'second_assists': 0,
+            'team': 'Huck and Hope School of Handling (Hope)',
+            'threw_drops': 0,
+            'throw_aways': 1,
+            'total_efficiency': 0.0
+        },
+        'Kristie Ellis': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 1,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'Huck and Hope School of Handling (Hope)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Marie-Ange Gravel': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 1.0,
+            'd_points_against': 0,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 2000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 2000,
+            'second_assists': 0,
+            'team': 'Attack',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Martin Cloake': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 1.0,
+            'd_points_against': 0,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 2000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 2000,
+            'second_assists': 0,
+            'team': 'Attack',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Nick Amlin': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 1,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 1.0,
+            'd_points_against': 0,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 1,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 12500,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 12500,
+            'second_assists': 0,
+            'team': 'Attack',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Patrick Kenzie': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 1,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 1,
+            'o_points_for': 0,
+            'pay': 1000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 1000,
+            'second_assists': 0,
+            'team': 'Huck and Hope School of Handling (Hope)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Ryan Briggs': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 1,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'Huck and Hope School of Handling (Hope)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Vanessa Mann': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 1,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'Huck and Hope School of Handling (Hope)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        }
+    },
+    'week': 0
+}
+
+snapshots['StatsSnapshotTests::test_drop 1'] = {
+    'stats': {
+        'Craig Anderson': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 1,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 1,
+            'o_points_for': 0,
+            'pay': -500,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': -500,
+            'second_assists': 0,
+            'team': 'Betty White (B-Dub)',
+            'threw_drops': 1,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Geofford Seaborn': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 1,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'Betty White (B-Dub)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Graham Brown': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 1,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'Betty White (B-Dub)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Jaime Boss': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 1,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 1,
+            'o_points_for': 0,
+            'pay': -4000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': -4000,
+            'second_assists': 0,
+            'team': 'Betty White (B-Dub)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Jeff Hunt': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 1.0,
+            'd_points_against': 0,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 2000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 2000,
+            'second_assists': 0,
+            'team': 'Katy Parity (back by popular demand)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Kindha Gorman': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 1.0,
+            'd_points_against': 0,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 2000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 2000,
+            'second_assists': 0,
+            'team': 'Katy Parity (back by popular demand)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Laura Chambers Storey': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 1,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 1.0,
+            'd_points_against': 0,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 1,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 12500,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 12500,
+            'second_assists': 0,
+            'team': 'Katy Parity (back by popular demand)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Marcus Bordage': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 1.0,
+            'd_points_against': 0,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 2000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 2000,
+            'second_assists': 0,
+            'team': 'Katy Parity (back by popular demand)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Morgan Howard': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 1.0,
+            'd_points_against': 0,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 10500,
+            'points_played': 1,
+            'pulls': 1,
+            'salary_per_point': 10500,
+            'second_assists': 1,
+            'team': 'Katy Parity (back by popular demand)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Patrick Kenzie': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 1,
+            'o_points_for': 0,
+            'pay': 500,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 500,
+            'second_assists': 0,
+            'team': 'Betty White (B-Dub)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Stacey Wowchuk': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 1,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'Betty White (B-Dub)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Steve Chow': {
+            'assists': 1,
+            'callahan': 0,
+            'catches': 1,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 1.0,
+            'd_points_against': 0,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 13000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 13000,
+            'second_assists': 0,
+            'team': 'Katy Parity (back by popular demand)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        }
+    },
+    'week': 0
+}
+
+snapshots['StatsSnapshotTests::test_half 1'] = {
+    'stats': {
+        'An Tran': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 9500,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 9500,
+            'second_assists': 1,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Andrea Proulx': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 1,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 1,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 11500,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 11500,
+            'second_assists': 0,
+            'team': 'Nautical Disaster - Man Overboard! (SOS)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Darlene Riley': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'David Townsend': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 1000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 1000,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Frederic Caron': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 1,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'Nautical Disaster - Man Overboard! (SOS)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Heather McCabe': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Jay Thor Turner': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'Nautical Disaster - Man Overboard! (SOS)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Jeremy Bryan': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'Nautical Disaster - Man Overboard! (SOS)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Jon Rowe': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 1,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 1,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 11500,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 11500,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Jonathan Pindur': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'Nautical Disaster - Man Overboard! (SOS)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Karen Kavanagh': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 1000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 1000,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Kevin Barford': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Kevin Hughes': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 1,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 10000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 10000,
+            'second_assists': 1,
+            'team': 'Nautical Disaster - Man Overboard! (SOS)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Matthew Schijns': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 1000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 1000,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Megan Robb': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 1000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 1000,
+            'second_assists': 0,
+            'team': 'Nautical Disaster - Man Overboard! (SOS)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        "Michael O'Hare": {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 1500,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 1500,
+            'second_assists': 0,
+            'team': 'Nautical Disaster - Man Overboard! (SOS)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Nicholas Aghajanian': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 1000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 1000,
+            'second_assists': 0,
+            'team': 'Nautical Disaster - Man Overboard! (SOS)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Owen Lumley': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 1,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Ryan Mussell': {
+            'assists': 1,
+            'callahan': 0,
+            'catches': 1,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 12000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 12000,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Sandra Hanson': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'Nautical Disaster - Man Overboard! (SOS)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Stephen Close': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Taka Yamada': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'Nautical Disaster - Man Overboard! (SOS)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Will Leckie': {
+            'assists': 1,
+            'callahan': 0,
+            'catches': 1,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 12000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 12000,
+            'second_assists': 0,
+            'team': 'Nautical Disaster - Man Overboard! (SOS)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Wing-Leung Chan': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        }
+    },
+    'week': 0
+}
+
+snapshots['StatsSnapshotTests::test_mini_game 1'] = {
+    'stats': {
+        'Ashlin Kelly': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 1000,
+            'points_played': 2,
+            'pulls': 0,
+            'salary_per_point': 500,
+            'second_assists': 0,
+            'team': 'Kells Angels Bicycle Club',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.5
+        },
+        'Benjamin Piper': {
+            'assists': 1,
+            'callahan': 0,
+            'catches': 2,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.5,
+            'o_points_against': 1,
+            'o_points_for': 1,
+            'pay': 11500,
+            'points_played': 2,
+            'pulls': 0,
+            'salary_per_point': 5750,
+            'second_assists': 0,
+            'team': 'Kells Angels Bicycle Club',
+            'threw_drops': 1,
+            'throw_aways': 0,
+            'total_efficiency': 0.5
+        },
+        'Brent Burton': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 1000,
+            'points_played': 2,
+            'pulls': 0,
+            'salary_per_point': 500,
+            'second_assists': 0,
+            'team': '99 Problems',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.5
+        },
+        'Brian Kells': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 9500,
+            'points_played': 2,
+            'pulls': 1,
+            'salary_per_point': 4750,
+            'second_assists': 1,
+            'team': 'Kells Angels Bicycle Club',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.5
+        },
+        'Carrie-Anne Whyte': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 1,
+            'goals': 0,
+            'o_efficiency': 0.5,
+            'o_points_against': 1,
+            'o_points_for': 1,
+            'pay': -3000,
+            'points_played': 2,
+            'pulls': 0,
+            'salary_per_point': -1500,
+            'second_assists': 0,
+            'team': 'Kells Angels Bicycle Club',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.5
+        },
+        'Chris Sullivan': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 0.5,
+            'd_points_against': 1,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 10500,
+            'points_played': 2,
+            'pulls': 0,
+            'salary_per_point': 5250,
+            'second_assists': 1,
+            'team': '99 Problems',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.5
+        },
+        'Christine Beals': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 1000,
+            'points_played': 2,
+            'pulls': 0,
+            'salary_per_point': 500,
+            'second_assists': 0,
+            'team': 'Kells Angels Bicycle Club',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.5
+        },
+        'Christopher Keates': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 2,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 0.5,
+            'd_points_against': 1,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 2,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 32500,
+            'points_played': 3,
+            'pulls': 0,
+            'salary_per_point': 10833,
+            'second_assists': 1,
+            'team': '99 Problems',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.6666666666666666
+        },
+        'Jason Fraser': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 1000,
+            'points_played': 2,
+            'pulls': 0,
+            'salary_per_point': 500,
+            'second_assists': 0,
+            'team': '99 Problems',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.5
+        },
+        'Jessie Robinson': {
+            'assists': 1,
+            'callahan': 0,
+            'catches': 1,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 0.5,
+            'd_points_against': 1,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 13000,
+            'points_played': 2,
+            'pulls': 0,
+            'salary_per_point': 6500,
+            'second_assists': 0,
+            'team': '99 Problems',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.5
+        },
+        'Jim Robinson': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 1,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 1,
+            'o_efficiency': 0.5,
+            'o_points_against': 1,
+            'o_points_for': 1,
+            'pay': 11500,
+            'points_played': 2,
+            'pulls': 0,
+            'salary_per_point': 5750,
+            'second_assists': 0,
+            'team': 'Kells Angels Bicycle Club',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.5
+        },
+        'Jonathan Champagne': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 1000,
+            'points_played': 2,
+            'pulls': 0,
+            'salary_per_point': 500,
+            'second_assists': 0,
+            'team': 'Kells Angels Bicycle Club',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.5
+        },
+        'Justine Price': {
+            'assists': 1,
+            'callahan': 0,
+            'catches': 1,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 12000,
+            'points_played': 2,
+            'pulls': 0,
+            'salary_per_point': 6000,
+            'second_assists': 0,
+            'team': '99 Problems',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.5
+        },
+        'Krys Kudakiewicz': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.5,
+            'd_points_against': 1,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 2000,
+            'points_played': 2,
+            'pulls': 0,
+            'salary_per_point': 1000,
+            'second_assists': 0,
+            'team': '99 Problems',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.5
+        },
+        'Martin Cloake': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 1,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 1,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 11500,
+            'points_played': 2,
+            'pulls': 0,
+            'salary_per_point': 5750,
+            'second_assists': 0,
+            'team': 'Kells Angels Bicycle Club',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.5
+        },
+        'Meagan Doyle': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 1000,
+            'points_played': 2,
+            'pulls': 0,
+            'salary_per_point': 500,
+            'second_assists': 0,
+            'team': '99 Problems',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.5
+        },
+        'Michelle Warren': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.5,
+            'd_points_against': 1,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 2000,
+            'points_played': 2,
+            'pulls': 0,
+            'salary_per_point': 1000,
+            'second_assists': 0,
+            'team': '99 Problems',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.5
+        },
+        'Patrick McKelvey': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 1,
+            'completions': 2,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.5,
+            'o_points_against': 1,
+            'o_points_for': 1,
+            'pay': 10500,
+            'points_played': 2,
+            'pulls': 0,
+            'salary_per_point': 5250,
+            'second_assists': 1,
+            'team': 'Kells Angels Bicycle Club',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.5
+        },
+        'Richard Gregory': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 1000,
+            'points_played': 2,
+            'pulls': 0,
+            'salary_per_point': 500,
+            'second_assists': 0,
+            'team': '99 Problems',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.5
+        },
+        'Rob Ives': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.5,
+            'd_points_against': 1,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 2000,
+            'points_played': 2,
+            'pulls': 0,
+            'salary_per_point': 1000,
+            'second_assists': 0,
+            'team': '99 Problems',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.5
+        },
+        'Sam Lee': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.5,
+            'o_points_against': 1,
+            'o_points_for': 1,
+            'pay': 1000,
+            'points_played': 2,
+            'pulls': 0,
+            'salary_per_point': 500,
+            'second_assists': 0,
+            'team': 'Kells Angels Bicycle Club',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.5
+        },
+        'Scott Higgins': {
+            'assists': 1,
+            'callahan': 0,
+            'catches': 1,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 12000,
+            'points_played': 2,
+            'pulls': 0,
+            'salary_per_point': 6000,
+            'second_assists': 0,
+            'team': 'Kells Angels Bicycle Club',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.5
+        },
+        'Simon Berry': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': '99 Problems',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Tim Kealey': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.5,
+            'o_points_against': 1,
+            'o_points_for': 1,
+            'pay': 1500,
+            'points_played': 2,
+            'pulls': 0,
+            'salary_per_point': 750,
+            'second_assists': 0,
+            'team': 'Kells Angels Bicycle Club',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.5
+        }
+    },
+    'week': 0
+}
+
+snapshots['StatsSnapshotTests::test_mini_game2 1'] = {
+    'stats': {
+        'An Tran': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Darlene Riley': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'David Townsend': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Heather McCabe': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 1000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 1000,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Jamie Wildgen': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 1,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'Soho',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Jon Rowe': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Karen Kavanagh': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 1000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 1000,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Kevin Barford': {
+            'assists': 1,
+            'callahan': 0,
+            'catches': 1,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 12000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 12000,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Kirsten Querbach': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 1,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 1,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 11500,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 11500,
+            'second_assists': 0,
+            'team': 'Soho',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Kyle Sprysa': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 1,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 10000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 10000,
+            'second_assists': 1,
+            'team': 'Soho',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Laura Knowles': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'Soho',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Matthew Schijns': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Melissa Jess': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'Soho',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Michael Colantonio': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'Soho',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Nina Ramic': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 1500,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 1500,
+            'second_assists': 0,
+            'team': 'Soho',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Owen Lumley': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 9500,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 9500,
+            'second_assists': 1,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Rob Tyson': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 1000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 1000,
+            'second_assists': 0,
+            'team': 'Soho',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Ryan Mussell': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Sebastien Belanger': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'Soho',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Sina Dee': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 1,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'Soho',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Stephen Close': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 1000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 1000,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Thuc Nguyen': {
+            'assists': 1,
+            'callahan': 0,
+            'catches': 1,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 12000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 12000,
+            'second_assists': 0,
+            'team': 'Soho',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Tyler Mulcock': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 1000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 1000,
+            'second_assists': 0,
+            'team': 'Soho',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Wing-Leung Chan': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 1,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 1,
+            'o_efficiency': 1.0,
+            'o_points_against': 0,
+            'o_points_for': 1,
+            'pay': 11500,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 11500,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        }
+    },
+    'week': 0
+}
+
+snapshots['StatsSnapshotTests::test_throw_away 1'] = {
+    'stats': {
+        'Amos Lee': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 1,
+            'o_points_for': 0,
+            'pay': 500,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 500,
+            'second_assists': 0,
+            'team': 'Ultra Tide: Don an unsullied hue',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Angela Mueller': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 1,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'Ultra Tide: Don an unsullied hue',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Greg Probe': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 1,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 1,
+            'o_points_for': 0,
+            'pay': -4500,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': -4500,
+            'second_assists': 0,
+            'team': 'Ultra Tide: Don an unsullied hue',
+            'threw_drops': 0,
+            'throw_aways': 1,
+            'total_efficiency': 0.0
+        },
+        'Jonathan Pindur': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 1.0,
+            'd_points_against': 0,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 2000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 2000,
+            'second_assists': 0,
+            'team': 'Nautical Disaster - Man Overboard! (SOS)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Kevin Hughes': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 1.0,
+            'd_points_against': 0,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 10500,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 10500,
+            'second_assists': 1,
+            'team': 'Nautical Disaster - Man Overboard! (SOS)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Mark Donahue': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 1,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 1,
+            'o_points_for': 0,
+            'pay': 1000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 1000,
+            'second_assists': 0,
+            'team': 'Ultra Tide: Don an unsullied hue',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Megan Robb': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 1.0,
+            'd_points_against': 0,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 2000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 2000,
+            'second_assists': 0,
+            'team': 'Nautical Disaster - Man Overboard! (SOS)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        "Michael O'Hare": {
+            'assists': 1,
+            'callahan': 0,
+            'catches': 1,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 1.0,
+            'd_points_against': 0,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 13000,
+            'points_played': 1,
+            'pulls': 1,
+            'salary_per_point': 13000,
+            'second_assists': 0,
+            'team': 'Nautical Disaster - Man Overboard! (SOS)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Paula Saliba': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 1,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'Ultra Tide: Don an unsullied hue',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Taka Yamada': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 1.0,
+            'd_points_against': 0,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 2000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 2000,
+            'second_assists': 0,
+            'team': 'Nautical Disaster - Man Overboard! (SOS)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Trevor Stocki': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 1,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'Ultra Tide: Don an unsullied hue',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Will Leckie': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 1,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 1.0,
+            'd_points_against': 0,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 1,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 12500,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 12500,
+            'second_assists': 0,
+            'team': 'Nautical Disaster - Man Overboard! (SOS)',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        }
+    },
+    'week': 0
+}
+
+snapshots['StatsSnapshotTests::test_turnovers 1'] = {
+    'stats': {
+        'Heather McCabe': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 1,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Jamie Wildgen': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 1.0,
+            'd_points_against': 0,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 2000,
+            'points_played': 1,
+            'pulls': 1,
+            'salary_per_point': 2000,
+            'second_assists': 0,
+            'team': 'Soho',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Karen Kavanagh': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 1,
+            'o_points_for': 0,
+            'pay': 0,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 0,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Kevin Barford': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 3,
+            'completions': 2,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 1,
+            'o_points_for': 0,
+            'pay': 1500,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 1500,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 1,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Laura Knowles': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 1.0,
+            'd_points_against': 0,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 2000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 2000,
+            'second_assists': 0,
+            'team': 'Soho',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Melissa Jess': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 1.0,
+            'd_points_against': 0,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 2000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 2000,
+            'second_assists': 0,
+            'team': 'Soho',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Michael Colantonio': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 2,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 1.0,
+            'd_points_against': 0,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 2500,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 2500,
+            'second_assists': 0,
+            'team': 'Soho',
+            'threw_drops': 1,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Owen Lumley': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 0,
+            'completions': 3,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 1,
+            'o_points_for': 0,
+            'pay': 1500,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 1500,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 0.0
+        },
+        'Sebastien Belanger': {
+            'assists': 0,
+            'callahan': 1,
+            'catches': 0,
+            'completions': 2,
+            'd_blocks': 0,
+            'd_efficiency': 1.0,
+            'd_points_against': 0,
+            'd_points_for': 1,
+            'drops': 0,
+            'goals': 1,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': 13000,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': 13000,
+            'second_assists': 0,
+            'team': 'Soho',
+            'threw_drops': 0,
+            'throw_aways': 0,
+            'total_efficiency': 1.0
+        },
+        'Sina Dee': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 1,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 1.0,
+            'd_points_against': 0,
+            'd_points_for': 1,
+            'drops': 1,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 0,
+            'o_points_for': 0,
+            'pay': -16500,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': -16500,
+            'second_assists': 0,
+            'team': 'Soho',
+            'threw_drops': 0,
+            'throw_aways': 3,
+            'total_efficiency': 1.0
+        },
+        'Stephen Close': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 1,
+            'completions': 0,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 0,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 1,
+            'o_points_for': 0,
+            'pay': -4500,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': -4500,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 1,
+            'total_efficiency': 0.0
+        },
+        'Wing-Leung Chan': {
+            'assists': 0,
+            'callahan': 0,
+            'catches': 2,
+            'completions': 1,
+            'd_blocks': 0,
+            'd_efficiency': 0.0,
+            'd_points_against': 0,
+            'd_points_for': 0,
+            'drops': 1,
+            'goals': 0,
+            'o_efficiency': 0.0,
+            'o_points_against': 1,
+            'o_points_for': 0,
+            'pay': -17500,
+            'points_played': 1,
+            'pulls': 0,
+            'salary_per_point': -17500,
+            'second_assists': 0,
+            'team': 'lumleysexuals',
+            'threw_drops': 0,
+            'throw_aways': 3,
+            'total_efficiency': 0.0
+        }
+    },
+    'week': 0
+}
